scripts/requirements.txt
geopandas==0.14.1
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23